OOPath tests - Adds changes from PR review
- Moves model into separate package
- Small other fixes
