+++ drools-compiler/src/test/java/org/drools/compiler/oopath/OOPathBenchmarkTest.java
@@ -15,21 +15,23 @@
 
 package org.drools.compiler.oopath;
 
+import java.util.ArrayList;
+import java.util.List;
+import org.assertj.core.api.Assertions;
+import org.drools.compiler.oopath.model.Child;
+import org.drools.compiler.oopath.model.Man;
+import org.drools.compiler.oopath.model.Toy;
+import org.drools.compiler.oopath.model.Woman;
 import org.drools.core.common.InternalFactHandle;
 import org.kie.api.KieBase;
 import org.kie.api.io.ResourceType;
 import org.kie.api.runtime.KieSession;
 import org.kie.internal.utils.KieHelper;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import static org.junit.Assert.assertEquals;
-
 public class OOPathBenchmarkTest {
 
     private static final String RELATIONAL_DRL =
-            "import org.drools.compiler.oopath.*;\n" +
+            "import org.drools.compiler.oopath.model.*;\n" +
             "global java.util.List list\n" +
             "\n" +
             "rule R when\n" +
@@ -42,7 +44,7 @@ public class OOPathBenchmarkTest {
             "end\n";
 
     private static final String FROM_DRL =
-            "import org.drools.compiler.oopath.*;\n" +
+            "import org.drools.compiler.oopath.model.*;\n" +
             "global java.util.List list\n" +
             "\n" +
             "rule R when\n" +
@@ -54,7 +56,7 @@ public class OOPathBenchmarkTest {
             "end\n";
 
     private static final String OOPATH_DRL =
-            "import org.drools.compiler.oopath.*;\n" +
+            "import org.drools.compiler.oopath.model.*;\n" +
             "global java.util.List list\n" +
             "\n" +
             "rule R when\n" +
@@ -166,7 +168,7 @@ public class OOPathBenchmarkTest {
         ksession.fireAllRules();
         result[1] = System.nanoTime() - start;
 
-        assertEquals(n, list.size());
+        Assertions.assertThat(n).isEqualTo(list.size());
         ksession.dispose();
 
         return result;
@@ -199,7 +201,7 @@ public class OOPathBenchmarkTest {
         ksession.fireAllRules();
         result[1] = System.nanoTime() - start;
 
-        assertEquals(n * 3, list.size());
+        Assertions.assertThat(n * 3).isEqualTo(list.size());
         ksession.dispose();
 
         return result;
@@ -229,7 +231,7 @@ public class OOPathBenchmarkTest {
         ksession.fireAllRules();
         result[1] = System.nanoTime() - start;
 
-        assertEquals(n, list.size());
+        Assertions.assertThat(n).isEqualTo(list.size());
         ksession.dispose();
 
         return result;
