--- drools-compiler/src/test/java/org/drools/compiler/oopath/Woman.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
-*/
-
-package org.drools.compiler.oopath;
-
-public class Woman extends Adult {
-
-    private String husband;
-
-    public Woman(String name, int age) {
-        super(name, age);
-    }
-
-    public String getHusband() {
-        return husband;
-    }
-
-    public void setHusband(String husband) {
-        this.husband = husband;
-    }
-}
