+++ drools-compiler/src/test/java/org/drools/compiler/oopath/model/Group.java
@@ -1,48 +1,47 @@
 /*
- * Copyright 2016 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
-*/
+ */
 
-package org.drools.compiler.oopath;
+package org.drools.compiler.oopath.model;
+
+import java.util.List;
 
 import org.drools.core.phreak.AbstractReactiveObject;
+import org.drools.core.phreak.ReactiveList;
 
-public class TMFile extends AbstractReactiveObject {
+public class Group extends AbstractReactiveObject {
     private final String name;
-    private int size;
+    private final List<Person> members = new ReactiveList<Person>();
 
-    public TMFile(String name, int size) {
+    public Group(String name) {
         this.name = name;
-        this.size = size;
     }
 
     public String getName() {
         return name;
     }
 
-    public int getSize() {
-        return size;
+    public List<Person> getMembers() {
+        return members;
     }
 
-    public void setSize(int size) {
-        this.size = size;
-        notifyModification();
+    public void addPerson(Person p) {
+        members.add(p);
     }
-
-    @Override
-    public String toString() {
-        StringBuilder builder = new StringBuilder();
-        builder.append("TMFile [name=").append(name).append(", size=").append(size).append("]");
-        return builder.toString();
+    
+    public void removePerson(Person p) {
+        members.remove(p);
     }
 }
