+++ drools-compiler/src/test/java/org/drools/compiler/oopath/model/Man.java
@@ -1,23 +1,34 @@
 /*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
-*/
+ */
 
-package org.drools.compiler.oopath;
+package org.drools.compiler.oopath.model;
 
-public class BabyBoy extends Child {
+public class Man extends Adult {
 
-    public BabyBoy(String name, int age) {
+    private Woman wife;
+
+    public Man(String name, int age) {
         super(name, age);
+    }
+
+    public Woman getWife() {
+        return wife;
+    }
+
+    public void setWife(Woman wife) {
+        this.wife = wife;
     }
 }
