+++ drools-compiler/src/test/java/org/drools/compiler/oopath/model/TMFileSet.java
@@ -1,32 +1,31 @@
 /*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2017 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
-*/
+ */
 
-package org.drools.compiler.oopath;
+package org.drools.compiler.oopath.model;
+
+import java.util.Set;
 
 import org.drools.core.phreak.AbstractReactiveObject;
-import org.drools.core.phreak.ReactiveList;
+import org.drools.core.phreak.ReactiveSet;
 
-import java.util.List;
-
-public class School extends AbstractReactiveObject {
-
+public class TMFileSet extends AbstractReactiveObject {
     private final String name;
+    private final Set<TMFile> members = new ReactiveSet<TMFile>();
 
-    private final List<Child> children = new ReactiveList<Child>();
-
-    public School(String name) {
+    public TMFileSet(String name) {
         this.name = name;
     }
 
@@ -34,11 +33,7 @@ public class School extends AbstractReactiveObject {
         return name;
     }
 
-    public List<Child> getChildren() {
-        return children;
-    }
-
-    public void addChild(Child child) {
-        children.add(child);
+    public Set<TMFile> getFiles() {
+        return members;
     }
 }
